tests: move all unit tests into unit folders
