+++ packages/web-app-files/tests/unit/helpers/resources.spec.js
@@ -1,4 +1,4 @@
-import { buildSharedResource } from '../../src/helpers/resources'
+import { buildSharedResource } from '../../../src/helpers/resources'
 
 describe('buildSharedResource', () => {
   const resourceFactory = (incomingShares = false) => {
