+++ packages/web-runtime/tests/unit/components/UserMenu.spec.js
@@ -2,7 +2,7 @@ import { shallowMount, createLocalVue, mount } from '@vue/test-utils'
 import Vuex from 'vuex'
 import { Store } from 'vuex-mock-store'
 import UserMenu from 'web-runtime/src/components/UserMenu.vue'
-import stubs from '../../../../tests/unit/stubs'
+import stubs from '../../../../../tests/unit/stubs'
 import DesignSystem from 'owncloud-design-system'
 
 const localVue = createLocalVue()
